Use peliasOneEdgeGram for all dynamic name fields
In Elasticsearch 2.0, fields with the same name in different _types can
no longer have different mappings. We currently use a one-gram mapping
on admin _types (country, locality, etc), to enable better autocomplete
results, but a two-gram mapping on other _types (address, venue, etc) to
reduce index size.

This won't work any more. Of the two, the peliasOneEdgeGram analyzer
will return good results in both cases, at the expense of a potentially
much larger index size. In the long run we should probably look at the
FST-based suggesters that are in Elasticsearch and in 2.x support better
geo filtering.

See the docs for more info on the change:
https://www.elastic.co/guide/en/elasticsearch/reference/2.0/breaking_20_mapping_changes.html#_conflicting_field_mappings