feat(ES5): Drop support for Elasticsearch 5
BREAKING CHANGE: This drops Elasticsearch 5 from the test matrix for this repo. While
it makes no other changes, it's a breaking change as it marks the point
where we stop supporting ES5.

This allows us to make changes that support only ES6/7 as we add support
for Elasticsearch 7.

Connects https://github.com/pelias/pelias/issues/831