chore(integration): Move schema creation options to common config
After looking at the changes in
https://github.com/pelias/schema/pull/394, I noticed that all the
parameters used to create test indices during the integration tests were
identical.

Because we will want to make changes to those options as part of our
move to support Elasticsearch 7, this PR is a pure refactoring that
moves those options to the `common` object passed to each integration
test.

This should allow us to make any future changes to index creation
options in a single place.