remove alpha ordinals from both queries and docs
- sometimes a street is in OSM as "twenty fifth street", which means if
  a user searches for it as "25th st", they won't find it
- sometimes a user will search for "twenty fifth street", but it's in
  OSM as "25th st"

this change takes care of both of those by synonymizing the "alpha
ordinal" to the bare number (25) in both cases above.

It does not lead to an index explosion because most streets are in OSM
as "NNth St", and this is not a bi-directional synonym.

This doesn't help with autocomplete, a user searching for "twent...."
will not find much of anything.

I tried very hard to get this to work with a multplexer filter of the
multiword + single-word filters and ... I just could not get it to work,
it would seemingly ignore one of the sub filters I specified no matter
what I did.