add a country_a2 field for iso2 codes
currently we only index iso3 codes in country_a, which is a problem in autocomplete for countries where the iso2 code isn't a prefix of the iso3, like China (CHN/CN) or Mexico (MEX/MX) meaning the query "beijing cn"  fails in autocomplete.

It's not a super common real world case for users (since they should hit the result by earlier) but does come up a fair bit in testing and looks embarasssing.

iso2 queries seem to work fine in geocoding, I think because slop lets us ignore them?

This will be paired with a change to pelias-model to automatically add iso2s.

Open to other approaches.