CI: Remove Node.js 12 from build matrix
We normally try to keep all our CI configurations identical, but the
schema repository is a bit of a special case since it has to do so many
extra things around Elasticsearch.

Apparently, one of those extra things, `npm install -g npm`, now
installs a new enough version of NPM that it doesn't work on Node.js 12.

We are removing Node.js 12 support globally soon anyway in
https://github.com/pelias/pelias/issues/950,
so we might as well remove it here to make all the tests pass again.