Remove JDK configuration from CI
For quite some time we have configured a specific version of the JDK in
our CI.

As far as I understand it, modern versions of Elasticsearch use a
bundled version of the JDK so this configuration may no longer be
relevant.

On a practical/security note: it also looks like [the
scripts](https://github.com/michaelklishin/jdk_switcher) used to switch
JDK versions are over 4 years old, and just calling a bash script from a
3rd party repository is a bit scary.

I don't see any changes due to the removal of this option, is it
possible we don't need it anymore?